% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bnr_datapaths}
\alias{bnr_datapaths}
\title{Table with default paths to commonly used spatial input files}
\format{
A \link{data.frame} containing paths to key spatial data sources.
}
\source{
Manually updated and curated by BNR researchers
}
\usage{
bnr_datapaths
}
\description{
This dataset contains path names to commonly-used spatial data files. Given that
those files are usually quite large, we here only describe where to find them internally
and not upload the data itself.
Medium-long-term this could be improved by relying on github LFS systems or our own gitlab
instance.
}
\details{
The file has the following columns:

\link{*} 'drive': The path to drive where the data is stored (Default: \code{'P:/bnr/'}). Can be system dependent (Windows/Linux).
\link{*} 'access': A non-structured field containing the list of people that have access (for example \code{'IBF'} or \code{'bnr'}).
\link{*} 'group': A field entry describing to what this file belongs to (i.e. \code{"EPIC"}).
\link{*} 'filename': The actual filename
}
\note{
To update or overwrite, load the file and update, then apply
\code{usethis::use_data(bnr_datapaths, overwrite = TRUE) }.
}
\keyword{datasets}
