% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_notIn.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Inverse of 'in' call}
\usage{
a \%notin\% b
}
\arguments{
\item{a}{First \code{\link{vector}} object.}

\item{b}{Second \code{\link{vector}} object.}
}
\value{
A \code{\link{vector}} of \code{\link{logical}} entries of \code{'a'} not present in \code{'b'}.
}
\description{
Calculates the set of entries not present in the second vector.
Added for convenience since this is not supported by default in R.
}
\examples{
# example code
}
\author{
Martin Jung
}
\keyword{utils}
