% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_emptyfolder.R
\name{misc_emptyfolder}
\alias{misc_emptyfolder}
\title{Check if a directory is empty (and create it if it doesn't exist)}
\usage{
misc_emptyfolder(dir_path)
}
\arguments{
\item{dir_path}{A character string specifying the path to the directory.}
}
\value{
A logical value: \code{TRUE} if the directory is empty (or newly created), \code{FALSE} otherwise.
}
\description{
This function checks whether a specified directory is empty.
If the directory does not exist, it will be created.
}
\details{
It also checks whether the directory is newly created or already existed.
}
\examples{
# Check and create a directory
dir_path <- tempfile()
is_empty <- misc_emptyfolder(dir_path)
print(is_empty)
unlink(dir_path, recursive = TRUE)
}
\author{
Martin Jung

Contributors: ChatGPT
}
