% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_sanitizeNames.R
\name{misc_sanitizeNames}
\alias{misc_sanitizeNames}
\title{Sanitize variable names}
\usage{
misc_sanitizeNames(names)
}
\arguments{
\item{names}{A \code{\link{vector}} of \code{\link{character}} vectors to be sanitized.}
}
\value{
A \code{\link{vector}} of sanitized \code{\link{character}}.
}
\description{
Prepared covariates often have special characters in their
variable names which can or can not be used in formulas or cause errors for
certain procedures. This function converts special characters (points, underscores or similar)
of variable names into a species format.
}
\examples{
# Correct variable names
vars <- c("Climate-temperature2015", "Elevation__sealevel", "Landuse.forest..meanshare")
misc_sanitizeNames(vars)

}
\concept{Inspired from [`inlabru`] \code{"bru_standardise_names"} function.}
\keyword{utils}
