% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_replaceGriddedNA.R
\name{sp_replaceGriddedNA}
\alias{sp_replaceGriddedNA}
\title{Replace NA values in gridded layers with a fixed value.}
\usage{
sp_replaceGriddedNA(obj, value = 0, mask, verbose = FALSE)
}
\arguments{
\item{obj}{A \code{\link{SpatRaster}}, \code{\link{SpatRasterDataset}} or \code{\link{stars}} object.}

\item{value}{A fixed numeric value of which all \code{NA} values are to be replaced with (Default: \code{0}).}

\item{mask}{An optional \code{\link{SpatRaster}} object used instead of the value.}

\item{verbose}{Be chatty about what is processed (Default: \code{FALSE}).}
}
\value{
A object of the same type as the input but with no-data values replaced with \code{0}.
}
\description{
This function replaces all NA values in a spatial gridded layer with
a fixed value such as for example \code{0}. Accepted input layers are for \code{\link{SpatRaster}} from
the \code{"terra"} R-package and \code{\link{stars}} from the \code{"stars"} R-package.
}
\details{
Required inputs are a single \code{"obj"} gridded data object and a numeric value.
In addition an optional mask layer can be provided that to use a mask. In this case
all no-data values a replaced with the value in this mask.
}
\examples{
# Example
s <- rast(system.file("ex/logo.tif", package="terra"))
s[]

}
\author{
Martin Jung
}
\keyword{internal,}
\keyword{utils}
