% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_exportNetCDF.R
\name{spl_exportNetCDF}
\alias{spl_exportNetCDF}
\title{RExport a gridded raster to a NetCDF format}
\usage{
spl_exportNetCDF(
  obj,
  filename,
  global_meta = system.file("iiasa_meta.yaml", package = "BNRTools"),
  separate_meta = FALSE,
  ...
)
}
\arguments{
\item{obj}{A \code{\link[terra:SpatRaster]{SpatRaster}} object to be exported.}

\item{filename}{A \code{\link{character}} with the output filename.}

\item{global_meta}{A global metadata descriptor by default using IIASA standard
metadata (Default: \code{"iiasa_meta data"}).}

\item{separate_meta}{A \code{\link{logical}} flag on whether the metadata should be
written separately in \code{"yaml"} format (Default: \code{FALSE}).}

\item{...}{Any other metadata that should be overwritten or added to \code{"global_meta"}.}
}
\description{
This function serves as a general wrapper function to export a provided
spatial gridded layer as multi-dimensional NetCDF file. It furthermore requires
the specification of a list containing metadata information.
}
\details{
The default metadata is contained in \code{"inst/iiasa_meta"}. See examples.
}
\note{
A support for \code{'stars'} could be added.
}
\examples{
\dontrun{
# Load default metadata (loaded by default by function too)
meta <- yaml::read_yaml(system.file("iiasa_meta.yaml", package = "BNRTools"))

# Dummy raster
obj <- terra::rast(ncol = 100, nrow = 100,
                   xmin = 0, xmax = 100,
                   ymin = 0, ymax = 100,
                   resolution = 5, crs = terra::crs("WGS84"),
                   val = runif(400)
                   )

# Export
spl_exportNetCDF(obj, filename = "test.nc",
                     global_meta = meta, title = "Super cool analysis")
}
}
\seealso{
\code{\link[terra]{writeCDF}},
\code{\link[terra]{writeRaster}}
}
\author{
Martin Jung
}
\keyword{spatial}
