% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_summarizeGPKGlayers.R
\name{spl_summarizeGPKGlayers}
\alias{spl_summarizeGPKGlayers}
\title{Summarize a set of layers from geopackages}
\usage{
spl_summarizeGPKGlayers(folder, verbose = TRUE)
}
\arguments{
\item{folder}{A \code{\link{character}} string specifying the path where the geopackage files are
stored. This looks specifically for files with the \code{'.gpkg'} extension, skipping others.}

\item{verbose}{A \code{\link{logical}} value indicating whether to print additional information during processing.}
}
\value{
A \code{\link{data.frame}} summarizing the layers in the geopackage, including their names, geometry types, and feature counts.
}
\description{
Geopackages are a common format for storing spatial data, allowing
multiple layers within a single file. These can be for example point or polygon
data.
}
\examples{
\dontrun{
 # Folder
 spl_summarizeGPKGlayers(folder = "path/to/your/geopackages")
}
}
\author{
Martin Jung
}
