% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_downscalr2ibis.R
\name{conv_downscalr2ibis}
\alias{conv_downscalr2ibis}
\title{Function to format a prepared GLOBIOM netCDF file for use in \code{ibis.iSDM}}
\usage{
conv_downscalr2ibis(
  fname,
  ignore = NULL,
  period = "all",
  template = NULL,
  shares_to_area = FALSE,
  use_gdalutils = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fname}{A filename in \code{\link{character}} pointing to a GLOBIOM output in netCDF format.}

\item{ignore}{A \code{\link{vector}} of variables to be ignored (Default: \code{NULL}).}

\item{period}{A \code{\link{character}} limiting the period to be returned from the
formatted data. Options include \code{"reference"} for the first entry, \code{"projection"}
for all entries but the first, and \code{"all"} for all entries (Default: \code{"reference"}).}

\item{template}{An optional \code{\link{SpatRaster}} object towards which projects
should be transformed.}

\item{shares_to_area}{A \code{\link{logical}} on whether shares should be corrected to
areas (if identified).}

\item{use_gdalutils}{(Deprecated) \code{\link{logical}} on to use gdalutils hack-around.}

\item{verbose}{\code{\link{logical}} on whether to be chatty.}
}
\value{
A \code{\link{SpatRaster}} stack with the formatted GLOBIOM predictors.
}
\description{
This function expects a downscaled GLOBIOM output as created in
the BIOCLIMA project. It converts the input to a stars object to be fed to
the \code{ibis.iSDM} R-package.
}
\examples{
\dontrun{
## Does not work unless downscalr file is provided.
# Expects a filename pointing to a netCDF file.
covariates <- conv_downscalr2ibis(fname)
}

}
\author{
Martin Jung
}
\keyword{conversion}
