% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_growGrid.R
\name{spl_growGrid}
\alias{spl_growGrid}
\title{Grow a categorical SpatRaster by certain amount of pixels or distance.}
\usage{
spl_growGrid(x, iter = 2)
}
\arguments{
\item{x}{A categorical \code{\link{SpatRaster}} to be grown.}

\item{iter}{A \code{\link{numeric}} values of the maximum number of iterations from the grid
cell border of which \code{x} should be grown (Default: \code{2}).}
}
\value{
\code{\link{SpatRaster}}
}
\description{
A common issue in aggregating categorical rasters is that coastal and boundary
gridcells tend to decrease owing to aggregation effects. This can cause issues
of non-matching grid cells later on.
This simple function takes a categorical \code{\link{SpatRaster}} object and grows it
into no-data areas (those with \code{NA} values) within an optionally provided
distance.
}
\note{
Only \code{NA} grid cells will be filled!
}
\examples{
set.seed(42)
ras <- terra::rast(ncol = 100, nrow = 100, xmin = 0, xmax = 100,
ymin = 0, ymax = 100, resolution = 10, crs = NA)

# Fill with dummy values but keep half as NA
terra::values(ras) <- c(rep(NA,50),rbinom(terra::ncell(ras)/2, 10, 0.5))

# Convert to factor
ras <- terra::as.factor(ras)
assertthat::assert_that(terra::is.factor(ras))

ras_nona <- spl_growGrid(x = ras, iter = 10)
terra::plot(ras_nona)

}
\seealso{
\code{\link[terra]{buffer}},
\code{\link[terra]{focal}}
}
\keyword{spatial}
