% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_mosaicTiffs.R
\name{spl_mosaicTiffs}
\alias{spl_mosaicTiffs}
\title{Mosaic a list of spatial layers together.}
\usage{
spl_mosaicTiffs(files, ofname = NULL, tempdir = NULL, dt = "INT2S", ...)
}
\arguments{
\item{files}{A \code{\link{character}} vector with filenames of spatial files.}

\item{ofname}{A \code{\link{character}} where the output should be written (Default: \code{NULL}).}

\item{tempdir}{A \code{\link{character}} with a temporary folder that must exist (Default: \code{NULL}).}

\item{dt}{A \code{\link{character}} with the output datatype of the spatial file (Default: \code{"INT2S"}).}

\item{...}{Any other parameters passed to \code{\link{writeRaster}}.}
}
\value{
A \code{\link{SpatRaster}} or just a file.
}
\description{
Particular for large spatial files, it is often impractical
to process the entire layer as single file. Instead such a file
could be split in smaller chunks and processed as such.

A common issue then is to reconstruct a spatial file of the original extent.
This function takes a list of filenames as input and mosaics them
together using a \code{'vrt'} file format.
}
\details{
This function by default uses the \code{'gdalUtilities'} R-package and tools
for most of the projections.
}
\examples{
\dontrun{
 # Get list of files
 ll <- list.files(path_to_folder)

 # Mosaic
 spl_mosaicTiffs(ll, "full_file.tif")
}

}
\seealso{
\code{\link[terra]{mosaic}},
\code{\link[gdalUtilities]{gdalbuildvrt}}
}
\author{
Martin Jung
}
\keyword{spatial}
