% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_resampleRas.R
\name{spl_resampleRas}
\alias{spl_resampleRas}
\title{Resample raster}
\usage{
spl_resampleRas(x, y, discrete = FALSE)
}
\arguments{
\item{x}{A \code{\link{SpatRaster}} to be resampled.}

\item{y}{A \code{\link{SpatRaster}} to which \code{x} will be resampled}

\item{discrete}{\code{\link{logical}} to specifiy if input raster has continitous or discrete values}
}
\value{
\code{\link{SpatRaster}}
}
\description{
Resample two raster to the spatial resolution using aggregation or disaggregation.
}
\examples{
set.seed(42)
ras_a <- terra::rast(ncol = 100, nrow = 100, xmin = 0, xmax = 100,
ymin = 0, ymax = 100, resolution = 20, crs = NA)

ras_b <- terra::rast(ncol = 100, nrow = 100, xmin = 0, xmax = 100,
ymin = 0, ymax = 100, resolution = 5, crs = NA)

 terra::values(ras_a) <- runif(n = terra::ncell(ras_a))
 terra::values(ras_b) <- runif(n = terra::ncell(ras_b))

 spl_resampleRas(x = ras_a, y = ras_b)

}
\seealso{
\code{\link[terra]{aggregate}},
\code{\link[terra]{disagg}}
}
\keyword{spatial}
