% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_rwr.R
\name{spl_rwr}
\alias{spl_rwr}
\title{Calculates a rarity-weighted richness estimate from modelled species distributions}
\usage{
spl_rwr(x, normalize = TRUE, column = NULL)
}
\arguments{
\item{x}{A \code{\link[terra:SpatRaster]{SpatRaster}} or alternatively \code{\link{data.frame}} object containing the modelled species distributions.}

\item{normalize}{A \code{\link{logical}} flag on whether to normalize the rarity-weighted richness ranks.}

\item{column}{An optional \code{\link{character}} value on whether the (Default: \code{NULL}).}
}
\value{
A logical value: \code{TRUE} if the directory is empty (or newly created), \code{FALSE} otherwise.
}
\description{
This function calculates a rarity-weighted richness estimate from modelled species distributions,
which can for example be obtained from the \code{`ibis.iSDM`} R-package. The input maps should ideally be binary
presence-absence maps, but the function can also handle continuous predictions.
}
\details{
The function
}
\examples{
\dontrun{
# Calculate rarity-weighted richness from a [`SpatRaster`].

}
}
\references{
\itemize{
\item Albuquerque, F., & Beier, P. (2016). Predicted rarity‐weighted richness, a new tool to prioritize sites for species representation. Ecology and Evolution, 6(22), 8107-8114.
\item Albuquerque, F., & Beier, P. (2015). Rarity-weighted richness: a simple and reliable alternative to integer programming and heuristic algorithms for minimum set and maximum coverage problems in conservation planning. PloS one, 10(3), e0119905.
}
}
\author{
Martin Jung
}
